Implement bulk processing of individual MAHs
After src.traverse module is executed on a series of halo IDS, it
appends their MAHs to the TSV file (in a long format). src.forge then
converts this ("casts") to a more useful (wide) data format, which can
be used by the src.plot to produce normalised (or not) MAH plot.